agent/requirements.txt
boto3
botocore
strands-agents
bedrock-agentcore
python-dotenv
requests